+++ src/pages/About/routes.js
@@ -1,0 +1,11 @@
+import About from '@/pages/About/About.vue'
+
+const routes = [
+  {
+    path: '/about',
+    name: 'About',
+    component: About
+  }
+]
+
+export default routes
