+++ src/pages/Event/Details.vue
