+++ src/pages/Event/Register.vue
