+++ src/router/index.js
@@ -1,47 +1,9 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import EventList from '../views/Event/List.vue'
-import EventLayout from '../views/Event/Layout.vue'
-import EventDetails from '../views/Event/Details.vue'
-import EventRegister from '../views/Event/Register.vue'
-import EventEdit from '../views/Event/Edit.vue'
-import About from '../views/About.vue'
 
-const routes = [
-  {
-    path: '/',
-    name: 'EventList',
-    component: EventList,
-    props: route => ({ page: parseInt(route.query.page) || 1 })
-  },
-  {
-    path: '/event/:id',
-    name: 'EventLayout',
-    props: true,
-    component: EventLayout,
-    children: [
-      {
-        path: '',
-        name: 'EventDetails',
-        component: EventDetails
-      },
-      {
-        path: 'register',
-        name: 'EventRegister',
-        component: EventRegister
-      },
-      {
-        path: 'edit',
-        name: 'EventEdit',
-        component: EventEdit
-      }
-    ]
-  },
-  {
-    path: '/about',
-    name: 'About',
-    component: About
-  }
-]
+import EventRoutes from '@/pages/Event/routes.js'
+import AboutRoutes from '@/pages/About/routes.js'
+
+const routes = [...EventRoutes, ...AboutRoutes]
 
 const router = createRouter({
   history: createWebHistory(process.env.BASE_URL),
